Removed abstractKeyring to implement IKeyring as an interface of Keyring classes
