+++ packages/caver-transaction/src/transactionTypes/abstractFeeDelegatedTransaction.js
@@ -24,7 +24,9 @@ const TransactionHasher = require('../transactionHasher/transactionHasher')
 const AbstractTransaction = require('./abstractTransaction')
 const { refineSignatures, typeDetectionFromRLPEncoding } = require('../transactionHelper/transactionHelper')
 const Keyring = require('../../../caver-wallet/src/keyring/keyringFactory')
-const AbstractKeyring = require('../../../caver-wallet/src/keyring/abstractKeyring')
+const SingleKeyring = require('../../../caver-wallet/src/keyring/singleKeyring')
+const MultipleKeyring = require('../../../caver-wallet/src/keyring/multipleKeyring')
+const RoleBasedKeyring = require('../../../caver-wallet/src/keyring/roleBasedKeyring')
 const { KEY_ROLE } = require('../../../caver-wallet/src/keyring/keyringHelper')
 const utils = require('../../../caver-utils/src')
 const SignatureData = require('../../../caver-wallet/src/keyring/signatureData')
@@ -93,7 +95,7 @@ class AbstractFeeDelegatedTransaction extends AbstractTransaction {
         if (_.isString(key)) {
             keyring = Keyring.createFromPrivateKey(key)
         }
-        if (!(keyring instanceof AbstractKeyring))
+        if (!(keyring instanceof SingleKeyring) && !(keyring instanceof MultipleKeyring) && !(keyring instanceof RoleBasedKeyring))
             throw new Error(
                 `Unsupported key type. The key parameter of the signAsFeePayer must be a single private key string, KlaytnWalletKey string, or Keyring instance.`
             )
