+++ packages/caver-wallet/src/keyring/roleBasedKeyring.js
@@ -17,7 +17,6 @@
 */
 
 const _ = require('lodash')
-const AbstractKeyring = require('./abstractKeyring')
 
 const utils = require('../../../caver-utils')
 const PrivateKey = require('./privateKey')
@@ -30,15 +29,28 @@ const { validateForSigning, validateIndexWithKeys, encryptKey, formatEncrypted }
  * representing a Keyring which includes `address` and `private keys` by roles.
  * @class
  */
-class RoleBasedKeyring extends AbstractKeyring {
+class RoleBasedKeyring {
     /**
      * creates a RoleBasedKeyring.
      * @param {string} address - The address of keyring.
      * @param {Array.<Array<string>>|Array.<Array<PrivateKey>>} keys - The keys to use in RoleBasedKeyring.
      */
     constructor(address, keys) {
-        super(address)
+        this.address = address
         this.keys = keys
+    }
+
+    /**
+     * @type {string}
+     */
+    get address() {
+        return this._address
+    }
+
+    set address(addressInput) {
+        if (!utils.isAddress(addressInput)) throw new Error(`Invalid address : ${addressInput}`)
+
+        this._address = utils.addHexPrefix(addressInput).toLowerCase()
     }
 
     /**
@@ -231,6 +243,35 @@ class RoleBasedKeyring extends AbstractKeyring {
         }
 
         return formatEncrypted(4, this._address, keyring, options)
+    }
+
+    /**
+     * returns KlaytnWalletKey format. If keyring uses more than one private key, this function will throw error.
+     *
+     * @return {string}
+     */
+    getKlaytnWalletKey() {
+        throw new Error(`Not supported for this class.`)
+    }
+
+    /**
+     * encrypts a keyring and returns a keystore v3 object.
+     *
+     * @param {string} password The password to be used for keyring encryption. The encrypted key store can be decrypted with this password.
+     * @param {object} options The options to use when encrypt a keyring. See `keyring.encrypt` for more detail about options.
+     * @return {object}
+     */
+    encryptV3(password, options) {
+        throw new Error(`Not supported for this class. Use 'keyring.encrypt(password)'.`)
+    }
+
+    /**
+     * returns true if keyring has decoupled key.
+     *
+     * @return {boolean}
+     */
+    isDecoupled() {
+        return true
     }
 }
